Implement coercing for different strides

+++ src/native/types.h
@@ -1,7 +1,10 @@
 #ifndef TYPE_H
 #define TYPE_H
 #include <stdbool.h>
-typedef unsigned long index_t;
+#include <stdint.h>
+
+// TODO : How to support 64 bit machines?
+typedef uint32_t index_t;
 typedef unsigned long long ull;
 
 #endif
